data_project/requirements.txt
Django==1.10.3
djangorestframework==3.5.3
numpy==1.11.2
pandas==0.19.1
python-dateutil==2.5.3
pytz==2016.7
rest-pandas==0.4.1
six==1.10.0
